+++ CHANGELOG.md
@@ -132,7 +132,7 @@
 - 更改签名私钥获取方式
 - Payment.list app[id] 参数改为必传
 - 红包去除多余的参数
-- batch_refund charges 字段格式修改
+- batch_refund payments 字段格式修改
 
 #### 修正
 - Webhooks 解析对象 batch_refund、batch_transfer、customs 对应事件错误
@@ -145,7 +145,7 @@
 - 签名生成规则变更
 - 列表查询接口换成 `list` 方法，代替 `all` 方法
 - 退款接口调用方式变更，直接使用 Refund 对象
-- Refund 对象添加 charge_order_no, transaction_no 属性
+- Refund 对象添加 payment_order_no, transaction_no 属性
 
 ## 2.1.9
 #### 修改